Move VisitController to visit package
